move PluginModal component to its own module
